maven: Refactor resolver integration tests into separate test projects
