+++ maven/bnd-resolver-maven-plugin/pom.xml
@@ -67,17 +67,4 @@
 		</dependency>
 	</dependencies>
 
-	<build>
-		<plugins>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-invoker-plugin</artifactId>
-				<configuration>
-					<properties>
-						<jsse.enableSNIExtension>false</jsse.enableSNIExtension>
-					</properties>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
 </project>
