+++ maven/bnd-resolver-maven-plugin/src/it/enroute/app.bndrun
