+++ maven/bnd-resolver-maven-plugin/src/it/enroute/debug.bndrun
