+++ maven/bnd-resolver-maven-plugin/src/it/enroute/inner-invoker.properties
@@ -1,0 +1,7 @@
+invoker.goals=--no-transfer-progress bnd-indexer:index@index bnd-indexer:index@test-index biz.aQute.bnd:bnd-resolver-maven-plugin:${bnd.version}:resolve
+
+# Run mvn with --debug for debug logging
+#invoker.debug=true
+
+# Run mvn in debugging mode and wait for a debugger to attach
+#invoker.environmentVariables.MAVEN_DEBUG_OPTS=-agentlib:jdwp=transport=dt_socket,server=y,suspend=y,address=8000
