+++ maven/bnd-resolver-maven-plugin/src/it/enroute/invoker.properties
