+++ maven/bnd-resolver-maven-plugin/src/it/index/index.xml.gz
