+++ maven/bnd-resolver-maven-plugin/src/it/index/org/apache/felix/org.apache.felix.eventadmin/1.4.6/org.apache.felix.eventadmin-1.4.6.jar
