+++ maven/bnd-resolver-maven-plugin/src/it/parent/pom.xml
@@ -7,25 +7,13 @@
 	<version>0.0.1</version>
 	<packaging>pom</packaging>
 
-	<modules>
-		<module>resolve</module>
-		<module>resolve-from-defaults</module>
-		<module>resolve-from-dependencies</module>
-		<module>resolve-from-dependencyManagement</module>
-		<module>resolve-from-scopes</module>
-		<module>resolve-from-inputbundles</module>
-		<module>enroute</module>
-		<module>resolve-infer-with-custom-bsn</module>
-		<module>resolve-with-project-props</module>
-	</modules>
-
 	<build>
 		<pluginManagement>
 			<plugins>
 				<plugin>
 					<groupId>org.apache.maven.plugins</groupId>
 					<artifactId>maven-compiler-plugin</artifactId>
-					<version>3.8.1</version>
+					<version>3.10.1</version>
 				</plugin>
 				<plugin>
 					<groupId>org.apache.maven.plugins</groupId>
@@ -35,12 +23,22 @@
 				<plugin>
 					<groupId>org.apache.maven.plugins</groupId>
 					<artifactId>maven-jar-plugin</artifactId>
-					<version>3.2.0</version>
+					<version>3.2.2</version>
+				</plugin>
+				<plugin>
+					<groupId>org.apache.maven.plugins</groupId>
+					<artifactId>maven-dependency-plugin</artifactId>
+					<version>3.3.0</version>
 				</plugin>
 				<plugin>
 					<groupId>org.apache.maven.plugins</groupId>
 					<artifactId>maven-invoker-plugin</artifactId>
-					<version>3.2.1</version>
+					<version>3.2.2</version>
+				</plugin>
+				<plugin>
+					<groupId>biz.aQute.bnd</groupId>
+					<artifactId>bnd-maven-plugin</artifactId>
+					<version>${bnd.version}</version>
 				</plugin>
 				<plugin>
 					<groupId>biz.aQute.bnd</groupId>
