+++ maven/bnd-resolver-maven-plugin/src/it/resolve-from-defaults/pom.xml
@@ -7,30 +7,15 @@
 		<groupId>biz.aQute.bnd-test</groupId>
 		<artifactId>resolver-test</artifactId>
 		<version>0.0.1</version>
+		<relativePath>../parent</relativePath>
 	</parent>
 
-	<artifactId>resolve-with-project-props</artifactId>
+	<artifactId>resolve-from-defaults</artifactId>
 	<version>0.0.1</version>
 
 	<properties>
 		<maven.compiler.target>1.8</maven.compiler.target>
-		<test.prop>false</test.prop>
 	</properties>
-
-	<dependencies>
-		<dependency>
-			<groupId>org.apache.felix</groupId>
-			<artifactId>org.apache.felix.eventadmin</artifactId>
-			<version>1.4.8</version>
-			<scope>runtime</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.felix</groupId>
-			<artifactId>org.apache.felix.framework</artifactId>
-			<version>5.4.0</version>
-			<scope>runtime</scope>
-		</dependency>
-	</dependencies>
 
 	<build>
 		<plugins>
@@ -48,12 +33,6 @@
 			<plugin>
 				<groupId>biz.aQute.bnd</groupId>
 				<artifactId>bnd-resolver-maven-plugin</artifactId>
-			</plugin>
-			<plugin>
-				<artifactId>maven-compiler-plugin</artifactId>
-				<configuration>
-					<target>${maven.compiler.target}</target>
-				</configuration>
 			</plugin>
 			<plugin>
 				<artifactId>maven-jar-plugin</artifactId>
