+++ maven/bnd-resolver-maven-plugin/src/it/resolve-from-defaults/test.bndrun
@@ -1,2 +1,2 @@
--standalone: ${.}/../index/index.xml.gz
+-standalone: ${projectsDirectory}/index/index.xml.gz
 -runfw: org.apache.felix.framework;version='[5.4.0,5.4.0]'
