+++ maven/bnd-resolver-maven-plugin/src/it/resolve-from-dependencies/pom.xml
@@ -7,20 +7,19 @@
 		<groupId>biz.aQute.bnd-test</groupId>
 		<artifactId>resolver-test</artifactId>
 		<version>0.0.1</version>
+		<relativePath>../parent</relativePath>
 	</parent>
 
-	<artifactId>resolve-infer-with-custom-bsn</artifactId>
+	<artifactId>resolve-from-dependencies</artifactId>
 	<version>0.0.1</version>
-
-	<properties>
-		<maven.compiler.target>1.8</maven.compiler.target>
-	</properties>
+	<packaging>pom</packaging>
 
 	<dependencies>
 		<dependency>
 			<groupId>org.apache.felix</groupId>
 			<artifactId>org.apache.felix.eventadmin</artifactId>
 			<version>1.4.8</version>
+			<scope>runtime</scope>
 		</dependency>
 		<dependency>
 			<groupId>org.apache.felix</groupId>
@@ -34,37 +33,7 @@
 		<plugins>
 			<plugin>
 				<groupId>biz.aQute.bnd</groupId>
-				<artifactId>bnd-maven-plugin</artifactId>
-				<configuration>
-					<bnd><![CDATA[
-						Bundle-SymbolicName: com.acme.foo
-					]]></bnd>
-				</configuration>
-				<executions>
-					<execution>
-						<goals>
-							<goal>bnd-process</goal>
-						</goals>
-					</execution>
-				</executions>
-			</plugin>
-			<plugin>
-				<groupId>biz.aQute.bnd</groupId>
 				<artifactId>bnd-resolver-maven-plugin</artifactId>
-			</plugin>
-			<plugin>
-				<artifactId>maven-compiler-plugin</artifactId>
-				<configuration>
-					<target>${maven.compiler.target}</target>
-				</configuration>
-			</plugin>
-			<plugin>
-				<artifactId>maven-jar-plugin</artifactId>
-				<configuration>
-					<archive>
-						<manifestFile>${project.build.outputDirectory}/META-INF/MANIFEST.MF</manifestFile>
-					</archive>
-				</configuration>
 			</plugin>
 		</plugins>
 	</build>
