+++ maven/bnd-resolver-maven-plugin/src/it/resolve-from-dependencies/test.bndrun
