+++ maven/bnd-resolver-maven-plugin/src/it/resolve-from-dependencyManagement/pom.xml
@@ -7,6 +7,7 @@
 		<groupId>biz.aQute.bnd-test</groupId>
 		<artifactId>resolver-test</artifactId>
 		<version>0.0.1</version>
+		<relativePath>../parent</relativePath>
 	</parent>
 
 	<artifactId>resolve-from-dependencyManagement</artifactId>
