+++ maven/bnd-resolver-maven-plugin/src/it/resolve-from-dependencyManagement/test.bndrun
