+++ maven/bnd-resolver-maven-plugin/src/it/resolve-from-inputbundles/bundles/org.apache.felix.eventadmin-1.4.8.jar
