+++ maven/bnd-resolver-maven-plugin/src/it/resolve-from-inputbundles/bundles/org.apache.felix.framework-5.4.0.jar
