+++ maven/bnd-resolver-maven-plugin/src/it/resolve-from-inputbundles/pom.xml
@@ -7,32 +7,26 @@
 		<groupId>biz.aQute.bnd-test</groupId>
 		<artifactId>resolver-test</artifactId>
 		<version>0.0.1</version>
+		<relativePath>../parent</relativePath>
 	</parent>
 
-	<artifactId>resolve-from-dependencies</artifactId>
+	<artifactId>resolve-from-inputbundles</artifactId>
 	<version>0.0.1</version>
 	<packaging>pom</packaging>
-
-	<dependencies>
-		<dependency>
-			<groupId>org.apache.felix</groupId>
-			<artifactId>org.apache.felix.eventadmin</artifactId>
-			<version>1.4.8</version>
-			<scope>runtime</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.felix</groupId>
-			<artifactId>org.apache.felix.framework</artifactId>
-			<version>5.4.0</version>
-			<scope>runtime</scope>
-		</dependency>
-	</dependencies>
 
 	<build>
 		<plugins>
 			<plugin>
 				<groupId>biz.aQute.bnd</groupId>
 				<artifactId>bnd-resolver-maven-plugin</artifactId>
+				<configuration>
+					<bundlesOnly>true</bundlesOnly>
+					<bundles>
+						<bundle>${project.basedir}/bundles/org.apache.felix.eventadmin-1.4.8.jar</bundle>
+						<bundle>${project.basedir}/bundles/org.apache.felix.framework-5.4.0.jar</bundle>
+					</bundles>
+					<useMavenDependencies>false</useMavenDependencies>
+				</configuration>
 			</plugin>
 		</plugins>
 	</build>
