+++ maven/bnd-resolver-maven-plugin/src/it/resolve-from-inputbundles/test.bndrun
