+++ maven/bnd-resolver-maven-plugin/src/it/resolve-from-scopes/test.bndrun
