+++ maven/bnd-resolver-maven-plugin/src/it/resolve-infer-with-custom-bsn/src/main/java/com/acme/Foo.java
