+++ maven/bnd-resolver-maven-plugin/src/it/resolve-infer-with-custom-bsn/test.bndrun
