+++ maven/bnd-resolver-maven-plugin/src/it/resolve-with-project-props/test.bndrun
