+++ maven/bnd-resolver-maven-plugin/src/it/resolve/pom.xml
@@ -7,25 +7,22 @@
 		<groupId>biz.aQute.bnd-test</groupId>
 		<artifactId>resolver-test</artifactId>
 		<version>0.0.1</version>
+		<relativePath>../parent</relativePath>
 	</parent>
 
-	<artifactId>resolve-from-inputbundles</artifactId>
+	<artifactId>resolve</artifactId>
 	<version>0.0.1</version>
 	<packaging>pom</packaging>
+
+	<properties>
+		<runrequires>osgi.identity;filter:='(osgi.identity=org.apache.felix.eventadmin)'</runrequires>
+	</properties>
 
 	<build>
 		<plugins>
 			<plugin>
 				<groupId>biz.aQute.bnd</groupId>
 				<artifactId>bnd-resolver-maven-plugin</artifactId>
-				<configuration>
-					<bundlesOnly>true</bundlesOnly>
-					<bundles>
-						<bundle>${project.basedir}/bundles/org.apache.felix.eventadmin-1.4.8.jar</bundle>
-						<bundle>${project.basedir}/bundles/org.apache.felix.framework-5.4.0.jar</bundle>
-					</bundles>
-					<useMavenDependencies>false</useMavenDependencies>
-				</configuration>
 			</plugin>
 		</plugins>
 	</build>
