--- maven/bnd-resolver-maven-plugin/src/it/test/resolve/pom.xml
@@ -1,28 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<parent>
-		<groupId>biz.aQute.bnd-test</groupId>
-		<artifactId>resolver-test</artifactId>
-		<version>0.0.1</version>
-	</parent>
-
-	<artifactId>resolve</artifactId>
-	<version>0.0.1</version>
-	<packaging>pom</packaging>
-
-	<properties>
-		<runrequires>osgi.identity;filter:='(osgi.identity=org.apache.felix.eventadmin)'</runrequires>
-	</properties>
-
-	<build>
-		<plugins>
-			<plugin>
-				<groupId>biz.aQute.bnd</groupId>
-				<artifactId>bnd-resolver-maven-plugin</artifactId>
-			</plugin>
-		</plugins>
-	</build>
-</project>
